+++ R/generate_ggplot.R
@@ -24,6 +24,7 @@
 #'    optional: legendtitle - Legend title, if NULL/not supplied, default is used \cr
 #'    optional: legendlocation - if "left" is specified, top left. Otherwise top. \cr
 #'    optional: linesize - Width of line, numeric, i.e. 1.5, 2, etc. set to 1.5 if not supplied. \cr
+#'    optional: pallette - overwrite plot colors by providing a vector of color names or hex numbers to be used for the plot. \cr
 #'    optional: title - A title for each plot. \cr
 #'    optional: for multiple plots, specify res[[1]]$ncols to define number of columns \cr
 #'
@@ -41,14 +42,18 @@
 generate_ggplot <- function(res)
 {
 
+    # change ggplot color palette to color-blind friendly
+    # http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/#a-colorblind-friendly-palette
+    # I added 3 more colors at the end to have 12, enough for all simulations
+    # the ones I added are likely not color-blind friendly but rarely used in the app
+    #cbfpalette <- c("#000000", "#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2", "#D55E00", "#CC79A7")  
+    cbfpalette <- c("#999999", "#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2", "#D55E00", "#CC79A7", "#00523B","#D5C711","#0019B2")
+    
     #nplots contains the number of plots to be produced.
     nplots = length(res) #length of list
 
     allplots=list() #will hold all plots
 
-    #lower and upper bounds for plots, these are used if none are provided by calling function
-    lb = 1e-10;
-    ub = 1e20;
 
     for (n in 1:nplots) #loop to create each plot
     {
@@ -95,7 +100,11 @@ generate_ggplot <- function(res)
       #see if user/calling function supplied x- and y-axis transformation information
       xscaletrans <- ifelse(is.null(resnow$xscale), 'identity',resnow$xscale)
       yscaletrans <- ifelse(is.null(resnow$yscale), 'identity',resnow$yscale)
-
+      
+      #lower and upper bounds for plots, these are used if none are provided by calling function
+      lb = 1e-10;
+      ub = 1e20;
+      
       #if we want a plot on log scale, set any value in the data at or below 0 to some small number
       if (xscaletrans !='identity') {dat$xvals[dat$xvals<=0]=lb}
       if (yscaletrans !='identity') {dat$yvals[dat$yvals<=0]=lb}
@@ -111,10 +120,11 @@ generate_ggplot <- function(res)
       #set line size as given by app or to 1.5 by default
       linesize = ifelse(is.null(resnow$linesize), 1.5, resnow$linesize)
 
-       #if the IDvar variable exists, use it for further stratification, otherwise just stratify on varnames
-	  if (is.null(dat$IDvar))
+    
+      #if the IDvar variable exists, use it for further stratification, otherwise just stratify on varnames
+	    if (is.null(dat$IDvar))
       {
-        p1 = ggplot2::ggplot(dat, ggplot2::aes(x = xvals, y = yvals, color = varnames, linetype = varnames, shape = varnames) )
+        p1 = ggplot2::ggplot(dat, ggplot2::aes(x = xvals, y = yvals, color = varnames, linetype = varnames, shape = varnames) )   
       }
       else
       {
@@ -124,21 +134,21 @@ generate_ggplot <- function(res)
       ###choose between different types of plots
       if (plottype == 'Scatterplot')
       {
-        p2 = p1 + ggplot2::geom_point( size = linesize, na.rm=TRUE)
+        p2 = p1 + ggplot2::geom_point( size = linesize, na.rm=TRUE) 
       }
       if (plottype == 'Boxplot')
       {
-        p2 = p1 + ggplot2::geom_boxplot(size = linesize, na.rm=TRUE)
+        p2 = p1 + ggplot2::geom_boxplot(size = linesize, na.rm=TRUE) 
       }
       if (plottype == 'Lineplot')
       {
-        p2 = p1 + ggplot2::geom_line(size = linesize, na.rm=TRUE)
+        p2 = p1 + ggplot2::geom_line(size = linesize, na.rm=TRUE) 
       }
       if (plottype == 'Mixedplot')
       {
         #a mix of lines and points. for this, the dataframe needs to contain an extra column indicating line or point
-        p1a = p1 + ggplot2::geom_line(data = dplyr::filter(dat,style == 'line'), size = linesize)
-        p2 = p1a + ggplot2::geom_point(data = dplyr::filter(dat,style == 'point'), size = 2.5*linesize)
+        p1a = p1 + ggplot2::geom_line(data = dplyr::filter(dat,style == 'line'), size = linesize) 
+        p2 = p1a + ggplot2::geom_point(data = dplyr::filter(dat,style == 'point'), size = 2.5*linesize) 
       }
 
 
@@ -147,7 +157,8 @@ generate_ggplot <- function(res)
 	 if (plottype == 'Boxplot')
       {
         p3 = p2 + ggplot2::scale_x_continuous(trans = xscaletrans, limits=c(xmin,xmax), breaks = NULL, labels = NULL)
-      }
+        p3 = p3 + ggplot2::labs(x = NULL)
+        }
       else
       {
         p3 = p2 + ggplot2::scale_x_continuous(trans = xscaletrans, limits=c(xmin,xmax))
@@ -165,9 +176,13 @@ generate_ggplot <- function(res)
       }
 
       #modify overall theme
-      p5 = p4 + ggplot2::theme_bw(base_size = 18)
+      p5 = p4 + ggplot2::theme_bw(base_size = 18) 
 
-
+      #default palette is set, overwritten if user provided
+      plotpalette = cbfpalette
+      if (!is.null(resnow$palette)) {plotpalette = resnow$palette }
+      
+            
       #do legend if TRUE or not provided
       if (is.null(resnow$makelegend) || resnow$makelegend)
       {
@@ -184,17 +199,18 @@ generate_ggplot <- function(res)
         p5a = p5 + ggplot2::theme(legend.key.width = grid::unit(3, "line"))
         p5b = p5a + ggplot2::theme(legend.position = legendlocation)
         p5c = p5b + ggplot2::scale_linetype_discrete(name = legendtitle) + ggplot2::scale_shape_discrete(name = legendtitle)
-        p5d = p5c + ggplot2::scale_colour_discrete(name = legendtitle)
-        p6 = p5d + guides(fill=guide_legend(title.position="top", nrow=3, byrow=TRUE))
+        p5d = p5c + ggplot2::scale_colour_manual(values=plotpalette, name = legendtitle) 
+        p6 = p5d + ggplot2::guides(fill=ggplot2::guide_legend(title.position="top", nrow=3, byrow=TRUE))
       }
       else
       {
-          p6 = p5 + ggplot2::theme(legend.position="none")
+          p6 = p5 + ggplot2::theme(legend.position="none") + ggplot2::scale_colour_manual(values=plotpalette) 
+          
       }
 
       #modify overall theme
       pfinal = p6
-
+      
       allplots[[n]] = pfinal
 
     } #end loop over individual plots