+++ NEWS.md
@@ -1,3 +1,12 @@
+# DSAIDE 0.8.2
+
+* Added new apps, updated a few others
+
+# DSAIDE 0.8.1
+
+* Fixed bug in environmental transmission app 
+* Updated readme.md and package vignette/tutorial
+
 # DSAIDE 0.8.0
 
 ## User facing changes